+++ src/sound/mod.rs
@@ -1,8 +1,13 @@
+#[cfg(feature = "sound")]
+mod default_sounds;
 #[cfg(not(feature = "sound"))]
 mod no_sound;
-#[cfg(feature = "sound")]
-mod play_sound;
+#[allow(clippy::module_inception)]
+mod sound;
 mod sound_config;
+#[cfg(feature = "sound")]
+mod sound_error;
+mod sound_library;
 #[cfg(feature = "sound")]
 mod sound_player;
 mod volume;
@@ -11,11 +16,14 @@ mod volume;
 pub use no_sound::*;
 #[cfg(feature = "sound")]
 pub use {
-    play_sound::*,
+    default_sounds::*,
+    sound_error::*,
     sound_player::*,
 };
 pub use {
+    sound::*,
     sound_config::*,
+    sound_library::*,
     volume::*,
 };