+++ src/sound/no_sound.rs
@@ -1,17 +1,25 @@
-use super::*;
+use {
+    super::*,
+    std::path::Path,
+};
 
 /// A dummy sound player which does nothing
 pub struct SoundPlayer {}
 impl SoundPlayer {
-    pub fn new(_base_volume: Volume) -> anyhow::Result<Self> {
+    pub fn new(
+        _base_volume: Volume,
+        _library: &SoundLibrary,
+        _base_dir: &Path,
+    ) -> anyhow::Result<Self> {
         Err(anyhow::anyhow!(
             "Bacon is compiled without the sound feature"
         ))
     }
     pub fn play(
         &self,
-        _beep: PlaySoundCommand,
-    ) {
+        _psc: &PlaySoundCommand,
+    ) -> anyhow::Result<()> {
         // should never be called as the  sound player is not instantiated
+        Ok(())
     }
 }